Remove Ingress and Cert conformance tests.
These have moved to knative.dev/networking and are being run in
downstream conformance testing. We will be running these in
knative/networking CI as well.
