+++ test/e2e-tests.sh
@@ -92,15 +92,8 @@ sleep 30
 # Run conformance and e2e tests.
 
 go_test_e2e -timeout=30m \
-  ./test/conformance/api/... ./test/conformance/runtime/... \
-  ./test/e2e \
+  ./test/conformance ./test/e2e \
   ${parallelism} \
-  "--resolvabledomain=$(use_resolvable_domain)" "${use_https}" "$(ingress_class)" || failed=1
-
-# We run KIngress conformance ingress separately, to make it easier to skip some tests.
-go_test_e2e -timeout=20m ./test/conformance/ingress ${parallelism}  \
-  `# Skip TestUpdate due to excessive flaking https://github.com/knative/serving/issues/8032` \
-  -run="TestIngressConformance/^[^u]" \
   "--resolvabledomain=$(use_resolvable_domain)" "${use_https}" "$(ingress_class)" || failed=1
 
 if (( HTTPS )); then
@@ -119,18 +112,6 @@ toggle_feature multi-container Disabled
 toggle_feature responsive-revision-gc Enabled
 go_test_e2e -timeout=2m ./test/e2e/gc || failed=1
 toggle_feature responsive-revision-gc Disabled
-
-# Certificate conformance tests must be run separately
-# because they need cert-manager specific configurations.
-kubectl apply -f ${TMP_DIR}/test/config/autotls/certmanager/selfsigned/
-add_trap "kubectl delete -f ${TMP_DIR}/test/config/autotls/certmanager/selfsigned/ --ignore-not-found" SIGKILL SIGTERM SIGQUIT
-go_test_e2e -timeout=10m ./test/conformance/certificate/nonhttp01 "$(certificate_class)" || failed=1
-kubectl delete -f ${TMP_DIR}/test/config/autotls/certmanager/selfsigned/
-
-kubectl apply -f ${TMP_DIR}/test/config/autotls/certmanager/http01/
-add_trap "kubectl delete -f ${TMP_DIR}/test/config/autotls/certmanager/http01/ --ignore-not-found" SIGKILL SIGTERM SIGQUIT
-go_test_e2e -timeout=10m ./test/conformance/certificate/http01 "$(certificate_class)" || failed=1
-kubectl delete -f ${TMP_DIR}/test/config/autotls/certmanager/http01/
 
 # Run scale tests.
 go_test_e2e -timeout=10m ${parallelism} ./test/scale || failed=1
